blupow_gateway/requirements-dev.txt
pytest
pytest-asyncio
black
isort
aiomqtt
paho-mqtt<2.0 